requirements.txt
torch>=2.4.1
transformers>=4.46.0
Pillow==10.4.0